requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.2.0
pytesseract>=0.3.10
Pillow>=10.0.0
chromadb>=0.4.0
anthropic>=0.18.0
numpy>=1.24.0